pyproject.toml
[project]
name = "little-milestones"
version = "0.1.0"
description = "Extract structured baby photo descriptions from local images."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "httpx>=0.28.1",
  "pillow>=11.2.1",
]

[project.optional-dependencies]
dev = [
  "piexif>=1.1.3",
  "pytest>=8.4.0",
]

[project.scripts]
littlems = "littlems.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/littlems"]

[tool.pytest.ini_options]
testpaths = ["tests"]
